backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-CORS==4.0.0
Flask-Mail==0.9.1
Werkzeug==2.3.7
celery==5.3.4
redis==5.0.1
python-dotenv==1.0.0
celery==5.3.4
flower==2.0.1
psutil==5.9.5